Add Script for Printing Live Home Project Structures
This script can be run on the `project.xml` file inside a Live Home 3D
Pro project, to see the structure of the stories, walls, etc. inside.
This is useful because the UI of LH3DP has a limit on how many
characters it will show in the "Project Tree" sidebar. It can be used to
identify any inconsistencies in the naming of objects before exporting
them to FBX for processing in Blender.