Fix Issues with Garbage that Starts with `_[0-9]{2}` after `Floor`
Live Home 3D Pro sometimes includes an underscore followed by digits in
the garbage text it appends to each object. This was causing floors to
be numbered randomly in the export. Now, we only support numbering after
`Floor` if the number is followed by `Slab`. In addition, ALL garbage
strings have to begin with an underscore because that appears to be the
convention from LH3D.